requirements.txt
https://github.com/ChantalMP/foolbox/blob/master/dist/foolbox-1.9.0-py3-none-any.whl?raw=true
numpy
keras==2.1.5
matplotlib
scikit-learn
Pillow
randomgen
torchvision
scipy==1.0.1
requests==2.20.0
opencv-python==3.4.0.12
scikit-image==0.13.1
tqdm==4.28.1
torch
torchvision
tensorflow-gpu==1.9
mypy
mem_top
pympler
sklearn
h5py